Enhance discount handling and add new migration
Enhanced discount handling and UI updates:
- Added `CashTenderAmount` to `Reset` in `TenderOrder.cs`.
- Introduced `GetElligiblePWDSCDiscount` in `OrderService.cs`.
- Updated `MainWindowViewModel.cs` to call the new method.
- Modified `AlternativePaymentsWindow.axaml.cs` to include `SaleTypeName`.
- Centered `InputStackPanel` and added `Submit_Button` in `GetSeniorPwdInfo.axaml`.
- Adjusted height of `SelectDiscountPwdScWindow.axaml`.
- Refactored `TenderOrderWindow.axaml` layout and added scroll viewer.
- Updated `TenderOrderWindow.axaml.cs` for new layout and receipt printing.
- Fixed UI state reset issue in `OrderSummaryView.axaml.cs`.
- General UI improvements for better user experience.

New migration for `EBISX_POS` project:
- Created tables: `AddOnType`, `Category`, `DrinkType`, `SaleType`, `User`, `Timestamp`, `AlternativePayments`, `CouponPromo`, `Menu`, `Order`, `Item`.
- Added indexes and foreign key constraints for data integrity.
- Defined `Down` method to reverse changes.

Other changes:
- Updated `JournalController.cs` with commented-out HTTP POST endpoints.
- Removed `20250404035811_EBISX_POS.Designer.cs` and `20250404035811_EBISX_POS.cs`.
- Changed `Id` type for `AlternativePayments` in `DataContextModelSnapshot.cs`.
- Added new methods to `IJournal` interface.
- Implemented journal entry methods in `JournalRepository.cs`.
- Updated `OrderRepository.cs` to call new journal methods.

+++ Migrations/DataContextModelSnapshot.cs
@@ -41,11 +41,11 @@ namespace EBISX_POS.API.Migrations
 
             modelBuilder.Entity("EBISX_POS.API.Models.AlternativePayments", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    MySqlPropertyBuilderExtensions.UseMySqlIdentityColumn(b.Property<int>("Id"));
+                    MySqlPropertyBuilderExtensions.UseMySqlIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<decimal>("Amount")
                         .HasColumnType("decimal(65,30)");
